Fixed "Nightmare", added botID config var.
The "Nightmare" command was broken because the response contained it's own triggering phrase, causing it to loop.
I fixed this by adding a "botID" config variable. You put the ID of the discord bot as the variable, and the bot ignores all messages from the account with that ID, aka itself.
I'm planning on making it find it's own ID automatically later, so you don't have to manually input it, but this is a decent solution for now.